📦 Struct: Move, Rename, Clean files.
