+++ Common.BasicHelper.Test/Core/Shell/Test_EnvironmentHelper.cs
@@ -1,0 +1,5 @@
+﻿namespace Common.BasicHelper.Test.Core.Shell;
+
+internal class Test_EnvironmentHelper
+{
+}
